+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project are documented in this file.
 - **native TLS**: Add `async-std-comp-native-tls` and `tokio-comp-native-tls` features ([#525](https://github.com/geofmureithi/apalis/pull/525))
 - **cron** : Introduce CronContext ([#488](https://github.com/geofmureithi/apalis/pull/488))
 - **stepped tasks** : adds ability to execute stepped tasks ([#478](https://github.com/geofmureithi/apalis/pull/478))
+- **SQL** : add support for job priority to SQL storages ([#533](https://github.com/geofmureithi/apalis/pull/533/))
 
 ### Fixed
 - **PostgresStorage**: PostgresStorage get_jobs status conditional ([#524](https://github.com/geofmureithi/apalis/pull/524))